_ARCHIVED/packages/concept-templates/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "iil-concept-templates"
version = "0.5.0"
description = "Shared schemas, frameworks and extraction for structured concept templates (ADR-147)"
readme = "README.md"
license = {text = "Proprietary"}
requires-python = ">=3.12"
authors = [
    { name = "Achim Dehnert", email = "achim.dehnert@iil.gmbh" },
]
dependencies = [
    "pydantic>=2.0",
]

[project.optional-dependencies]
pdf = ["pdfplumber>=0.10"]
render = ["weasyprint>=60.0"]
llm = [
    "iil-promptfw>=0.6.0",
    "iil-aifw>=0.5.0",
]
django = ["Django>=5.0,<6.0"]
knowledge = [
    "iil-outlinefw[knowledge]>=0.2.0",
]
full = [
    "pdfplumber>=0.10",
    "weasyprint>=60.0",
    "iil-promptfw>=0.6.0",
    "iil-aifw>=0.5.0",
    "Django>=5.0,<6.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[tool.hatch.build.targets.wheel]
packages = ["concept_templates"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
